+++ device/dev_common.go
@@ -4,20 +4,17 @@ import (
 	"bufio"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"os"
 	"os/exec"
+	"runtime"
 	"slices"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/danielpaulus/go-ios/ios"
-	"github.com/danielpaulus/go-ios/ios/imagemounter"
 	"github.com/shamanec/GADS-devices-provider/util"
 	log "github.com/sirupsen/logrus"
 )
@@ -27,6 +24,35 @@ var netClient = &http.Client{
 	Timeout: time.Second * 120,
 }
 var localDevices []*LocalDevice
+
+var DeviceMap = make(map[string]*LocalDevice)
+
+func UpdateDevices() {
+	Setup()
+
+	if runtime.GOOS == "linux" {
+		go updateDevicesLinux()
+	} else if runtime.GOOS == "darwin" {
+		go updateDevicesOSX()
+	} else if runtime.GOOS == "windows" {
+		go updateDevicesWindows()
+	}
+	go updateDevicesMongo()
+}
+
+// Create Mongo collections for all devices for logging
+// Create a map of *device.LocalDevice for easier access across the code
+func Setup() {
+	getLocalDevices()
+	createMongoLogCollectionsForAllDevices()
+	createDeviceMap()
+}
+
+func createDeviceMap() {
+	for _, device := range localDevices {
+		DeviceMap[device.Device.UDID] = device
+	}
+}
 
 // DEVICES SETUP
 
@@ -234,53 +260,6 @@ func getConnectedDevicesAndroid() []string {
 	return connectedDevices
 }
 
-// Check if xcodebuild is available on the host by checking its version
-func xcodebuildAvailable() bool {
-	cmd := exec.Command("xcodebuild", "-version")
-	log.WithFields(log.Fields{
-		"event": "provider",
-	}).Debug("Checking if xcodebuild is available on host")
-
-	if err := cmd.Run(); err != nil {
-		log.WithFields(log.Fields{
-			"event": "provider",
-		}).Warn("xcodebuild is not available or command failed - " + err.Error())
-		return false
-	}
-	return true
-}
-
-// Check if adb is available on the host by starting the server
-func adbAvailable() bool {
-	cmd := exec.Command("adb", "start-server")
-	log.WithFields(log.Fields{
-		"event": "provider",
-	}).Debug("Checking if adb is available on host")
-
-	if err := cmd.Run(); err != nil {
-		log.WithFields(log.Fields{
-			"event": "provider",
-		}).Warn("adb is not available or command failed - " + err.Error())
-		return false
-	}
-	return true
-}
-
-func goIOSAvailable() bool {
-	cmd := exec.Command("ios", "-h")
-	log.WithFields(log.Fields{
-		"event": "provider",
-	}).Debug("Checking if go-ios is available on host")
-
-	if err := cmd.Run(); err != nil {
-		log.WithFields(log.Fields{
-			"event": "provider",
-		}).Warn("go-ios is not available or command failed - " + err.Error())
-		return false
-	}
-	return true
-}
-
 func androidDevicesInConfig() bool {
 	for _, device := range localDevices {
 		if device.Device.OS == "android" {
@@ -317,304 +296,15 @@ func (device *LocalDevice) setContext() {
 	device.Context = ctx
 }
 
-// ANDROID DEVICES
-
-// Remove all adb forwarded ports(if any) on provider start
-func removeAdbForwardedPorts() {
-	util.ProviderLogger.LogDebug("provider", "Attempting to remove all `adb` forwarded ports on provider start")
-
-	cmd := exec.Command("adb", "forward", "--remove-all")
-	err := cmd.Run()
-	if err != nil {
-		util.ProviderLogger.LogWarn("provider", "Could not remove `adb` forwarded ports, there was an error or no devices are connected - "+err.Error())
-	}
-}
-
-func (device *LocalDevice) isGadsStreamServiceRunning() (bool, error) {
-	cmd := exec.CommandContext(device.Context, "adb", "-s", device.Device.UDID, "shell", "dumpsys", "activity", "services", "com.shamanec.stream/.ScreenCaptureService")
-	util.ProviderLogger.LogDebug("android_device_setup", fmt.Sprintf("Checking if GADS-stream is already running on Android device - %v", device.Device.UDID))
-
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return false, err
-	}
-
-	// If command returned "(nothing)" then the service is not running
-	if strings.Contains(string(output), "(nothing)") {
-		return false, nil
-	}
-
-	return true, nil
-}
-
-// Install gads-stream.apk on the device
-func (device *LocalDevice) installGadsStream() {
-	cmd := exec.CommandContext(device.Context, "adb", "-s", device.Device.UDID, "install", "-r", "./apps/gads-stream.apk")
-	util.ProviderLogger.LogDebug("android_device_setup", fmt.Sprintf("Installing GADS-stream apk on Android device - %v", device.Device.UDID))
-
-	err := cmd.Run()
-	if err != nil {
-		util.ProviderLogger.LogError("android_device_setup", fmt.Sprintf("Could not install GADS-stream on Android device - %v:\n %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-}
-
-// Add recording permissions to gads-stream app to avoid popup on start
-func (device *LocalDevice) addGadsStreamRecordingPermissions() {
-	cmd := exec.CommandContext(device.Context, "adb", "-s", device.Device.UDID, "shell", "appops", "set", "com.shamanec.stream", "PROJECT_MEDIA", "allow")
-	util.ProviderLogger.LogDebug("android_device_setup", fmt.Sprintf("Adding GADS-stream recording permissions on Android device - %v", device.Device.UDID))
-
-	err := cmd.Run()
-	if err != nil {
-		util.ProviderLogger.LogError("android_device_setup", fmt.Sprintf("Could not set GADS-stream recording permissions on Android device - %v:\n %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-}
-
-// Start the gads-stream app using adb
-func (device *LocalDevice) startGadsStreamApp() {
-	cmd := exec.CommandContext(device.Context, "adb", "-s", device.Device.UDID, "shell", "am", "start", "-n", "com.shamanec.stream/com.shamanec.stream.ScreenCaptureActivity")
-	util.ProviderLogger.LogDebug("android_device_setup", fmt.Sprintf("Starting GADS-stream app on Android device - %v - with command - `%v`", device.Device.UDID, cmd.Path))
-
-	err := cmd.Run()
-	if err != nil {
-		util.ProviderLogger.LogError("android_device_setup", fmt.Sprintf("Could not start GADS-stream app on Android device - %v:\n %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-}
-
-// Press the Home button using adb to hide the transparent gads-stream activity
-func (device *LocalDevice) pressHomeButton() {
-	cmd := exec.CommandContext(device.Context, "adb", "-s", device.Device.UDID, "shell", "input", "keyevent", "KEYCODE_HOME")
-	util.ProviderLogger.LogDebug("android_device_setup", fmt.Sprintf("Pressing Home button with adb on Android device - %v", device.Device.UDID))
-
-	err := cmd.Run()
-	if err != nil {
-		util.ProviderLogger.LogError("android_device_setup", fmt.Sprintf("Could not 'press' Home button with `adb` on Android device - %v, you need to press it yourself to hide the transparent activity og GADS-stream:\n %v", device.Device.UDID, err))
-	}
-}
-
-// Forward gads-stream socket to the host container
-func (device *LocalDevice) forwardGadsStream() {
-	cmd := exec.CommandContext(device.Context, "adb", "-s", device.Device.UDID, "forward", "tcp:"+device.Device.StreamPort, "tcp:1991")
-	util.ProviderLogger.LogDebug("android_device_setup", fmt.Sprintf("Forwarding GADS-stream port to host port %v for Android device - %v", device.Device.StreamPort, device.Device.UDID))
-
-	err := cmd.Run()
-	if err != nil {
-		util.ProviderLogger.LogError("android_device_setup", fmt.Sprintf("Could not forward GADS-stream port to host port %v for Android device - %v:\n %v", device.Device.StreamPort, device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-}
-
-// IOS DEVICES
-
-// Forward iOS device ports using `go-ios` CLI, for some reason using the library doesn't work properly
-func (device *LocalDevice) goIOSForward(hostPort string, devicePort string) {
-	cmd := exec.CommandContext(device.Context, "ios", "forward", hostPort, devicePort, "--udid="+device.Device.UDID)
-
-	// Create a pipe to capture the command's output
-	_, err := cmd.StdoutPipe()
-	if err != nil {
-		util.ProviderLogger.LogError("ios_device_setup", fmt.Sprintf("Could not create stdoutpipe executing `ios forward` for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	// Start the port forward command
-	err = cmd.Start()
-	if err != nil {
-		util.ProviderLogger.LogError("ios_device_setup", fmt.Sprintf("Error executing `ios forward` for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	if err := cmd.Wait(); err != nil {
-		util.ProviderLogger.LogError("ios_device_setup", fmt.Sprintf("Error waiting `ios forward` to finish for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-}
-
-func buildWebDriverAgent() error {
-	// Command to run continuously (replace with your command)
-	cmd := exec.Command("xcodebuild", "-project", "WebDriverAgent.xcodeproj", "-scheme", "WebDriverAgentRunner", "-destination", "generic/platform=iOS", "build-for-testing")
-	cmd.Dir = util.Config.EnvConfig.WDAPath
-
-	cmd.Stderr = os.Stderr
-	// Create a pipe to capture the command's output
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return err
-	}
-
-	util.ProviderLogger.LogInfo("provider", fmt.Sprintf("Starting WebDriverAgent xcodebuild in path `%s` with command `%s` ", util.Config.EnvConfig.WDAPath, cmd.String()))
-	if err := cmd.Start(); err != nil {
-		return err
-	}
-
-	// Create a scanner to read the command's output line by line
-	scanner := bufio.NewScanner(stdout)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		util.ProviderLogger.LogDebug("webdriveragent_xcodebuild", line)
-	}
-
-	// Wait for the command to finish
-	if err := cmd.Wait(); err != nil {
-		util.ProviderLogger.LogError("provider", fmt.Sprintf("Error waiting for build WebDriverAgent with `xcodebuild` command to finish - %s", err))
-		util.ProviderLogger.LogError("provider", "Building WebDriverAgent for testing was unsuccessful")
-		os.Exit(1)
-	}
-	return nil
-}
-
-func (device *LocalDevice) startWdaWithXcodebuild() {
-	// Create a usbmuxd.log file for Stderr
-	logger, _ := util.CreateCustomLogger("./logs/device_"+device.Device.UDID+"/wda.log", device.Device.UDID)
-
-	// Command to run continuously (replace with your command)
-	cmd := exec.CommandContext(device.Context, "xcodebuild", "-project", "WebDriverAgent.xcodeproj", "-scheme", "WebDriverAgentRunner", "-destination", "platform=iOS,id="+device.Device.UDID, "test-without-building", "-allowProvisioningUpdates")
-	cmd.Dir = util.Config.EnvConfig.WDAPath
-
-	// Create a pipe to capture the command's output
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		logger.LogError("webdriveragent_xcodebuild", fmt.Sprintf("Error creating stdoutpipe while running WebDriverAgent with xcodebuild for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	if err := cmd.Start(); err != nil {
-		logger.LogError("webdriveragent_xcodebuild", fmt.Sprintf("Could not start WebDriverAgent with xcodebuild for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	// Create a scanner to read the command's output line by line
-	scanner := bufio.NewScanner(stdout)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		logger.LogInfo("webdriveragent", strings.TrimSpace(line))
-
-		if strings.Contains(line, "Restarting after") {
-			device.resetLocalDevice()
-			return
-		}
-
-		if strings.Contains(line, "ServerURLHere") {
-			// device.DeviceIP = strings.Split(strings.Split(line, "//")[1], ":")[0]
-			device.WdaReadyChan <- true
-		}
-	}
-
-	if err := cmd.Wait(); err != nil {
-		logger.LogError("webdriveragent_xcodebuild", fmt.Sprintf("Error waiting for WebDriverAgent xcodebuild command to finish, it errored out or device `%v` was disconnected - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-}
-
-func (device *LocalDevice) getGoIOSDevice() {
-	goIosDevice, err := ios.GetDevice(device.Device.UDID)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "ios_device_setup",
-		}).Error(fmt.Sprintf("Could not get `go-ios` DeviceEntry for device - %v, err - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-
-	device.GoIOSDeviceEntry = goIosDevice
-}
-
 // HEALTH
-
-// Create a new WebDriverAgent session and update stream settings
-func (device *LocalDevice) updateWebDriverAgent() error {
-	util.ProviderLogger.LogInfo("ios_device_setup", fmt.Sprintf("Updating WebDriverAgent session and mjpeg stream settings for device `%s`", device.Device.UDID))
-
-	err := device.createWebDriverAgentSession()
-	if err != nil {
-		util.ProviderLogger.LogError("ios_device_setup", fmt.Sprintf("Could not create WebDriverAgent session for device %v - %v", device.Device.UDID, err))
-		return err
-	}
-
-	err = device.updateWebDriverAgentStreamSettings()
-	if err != nil {
-		util.ProviderLogger.LogError("ios_device_setup", fmt.Sprintf("Could not update WebDriverAgent stream settings for device %v - %v", device.Device.UDID, err))
-		return err
-	}
-
-	return nil
-}
-
-func (device *LocalDevice) updateWebDriverAgentStreamSettings() error {
-	// Set 30 frames per second, without any scaling, half the original screenshot quality
-	// TODO should make this configurable in some way, although can be easily updated the same way
-	requestString := `{"settings": {"mjpegServerFramerate": 30, "mjpegServerScreenshotQuality": 30, "mjpegScalingFactor": 100}}`
-
-	// Post the mjpeg server settings
-	response, err := http.Post("http://localhost:"+device.Device.WDAPort+"/session/"+device.Device.WDASessionID+"/appium/settings", "application/json", strings.NewReader(requestString))
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != 200 {
-		return errors.New("Could not successfully update WDA stream settings, status code=" + strconv.Itoa(response.StatusCode))
-	}
-
-	return nil
-}
-
-// Create a new WebDriverAgent session
-func (device *LocalDevice) createWebDriverAgentSession() error {
-	// TODO see if this JSON can be simplified
-	requestString := `{
-		"capabilities": {
-			"firstMatch": [{}],
-			"alwaysMatch": {
-				
-			}
-		}
-	}`
-
-	req, err := http.NewRequest(http.MethodPost, "http://localhost:"+device.Device.WDAPort+"/session", strings.NewReader(requestString))
-	if err != nil {
-		return err
-	}
-
-	response, err := netClient.Do(req)
-	if err != nil {
-		return err
-	}
-
-	// Get the response into a byte slice
-	responseBody, _ := io.ReadAll(response.Body)
-	// Unmarshal response into a basic map
-	var responseJson map[string]interface{}
-	err = json.Unmarshal(responseBody, &responseJson)
-	if err != nil {
-		return err
-	}
-
-	// Check the session ID from the map
-	if responseJson["sessionId"] == "" {
-		if err != nil {
-			return errors.New("could not get `sessionId` while creating a new WebDriverAgent session")
-		}
-	}
-
-	device.Device.WDASessionID = fmt.Sprintf("%v", responseJson["sessionId"])
-	return nil
-}
 
 // Loops checking if the Appium/WebDriverAgent servers for the device are alive and updates the DB each time
 func (device *LocalDevice) updateDeviceHealthStatus() {
 	ticker := time.NewTicker(1 * time.Second)
 	defer ticker.Stop()
 
-	util.ProviderLogger.LogInfo("device_setup", fmt.Sprintf("Started health status check for device %v. Will poll Appium/WebDriverAgent servers respective to the device each second", device.Device.UDID))
+	util.ProviderLogger.LogInfo("device_setup", fmt.Sprintf("Started health status check for device `%v`", device.Device.UDID))
+
 	for {
 		select {
 		case <-ticker.C:
@@ -657,27 +347,6 @@ func (device *LocalDevice) checkDeviceHealthStatus() {
 
 	device.Device.LastHealthyTimestamp = time.Now().UnixMilli()
 	device.Device.Healthy = true
-}
-
-// Check if the WebDriverAgent server for an iOS device is up
-func (device *LocalDevice) isWdaHealthy() (bool, error) {
-	req, err := http.NewRequest(http.MethodGet, "http://localhost:"+device.Device.WDAPort+"/status", nil)
-	if err != nil {
-		return false, err
-	}
-
-	response, err := netClient.Do(req)
-	if err != nil {
-		return false, err
-	}
-	defer response.Body.Close()
-
-	responseCode := response.StatusCode
-	if responseCode != 200 {
-		return false, nil
-	}
-
-	return true, nil
 }
 
 // APPIUM
@@ -776,98 +445,4 @@ func (device *LocalDevice) startAppium() {
 		util.ProviderLogger.LogError("device_setup", fmt.Sprintf("Error waiting for Appium command to finish, it errored out or device `%v` was disconnected - %v", device.Device.UDID, err))
 		device.resetLocalDevice()
 	}
-}
-
-func (device *LocalDevice) startWdaWithGoIOS() {
-	// Create a usbmuxd.log file for Stderr
-	wdaLogger, _ := util.CreateCustomLogger("./logs/device_"+device.Device.UDID+"/wda.log", device.Device.UDID)
-
-	cmd := exec.CommandContext(context.Background(), "ios", "runwda", "--bundleid="+util.Config.EnvConfig.WDABundleID, "--testrunnerbundleid="+util.Config.EnvConfig.WDABundleID, "--xctestconfig=WebDriverAgentRunner.xctest", "--udid="+device.Device.UDID)
-
-	fmt.Println("COMMAND IS: " + cmd.String())
-	// Create a pipe to capture the command's output
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		util.ProviderLogger.LogError("device_setup", fmt.Sprintf("Error creating stdoutpipe while running WebDriverAgent with go-ios for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	// Create a pipe to capture the command's error output
-	stderr, err := cmd.StderrPipe()
-	if err != nil {
-		util.ProviderLogger.LogError("device_setup", fmt.Sprintf("Error creating stderrpipe while running WebDriverAgent with go-ios for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	if err := cmd.Start(); err != nil {
-		util.ProviderLogger.LogError("device_setup", fmt.Sprintf("Could not start WebDriverAgent with go-ios for device `%v` - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-		return
-	}
-
-	// Create a combined reader from stdout and stderr
-	combinedReader := io.MultiReader(stderr, stdout)
-	// Create a scanner to read the command's output line by line
-	scanner := bufio.NewScanner(combinedReader)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		wdaLogger.LogInfo("webdriveragent", strings.TrimSpace(line))
-
-		if strings.Contains(line, "ServerURLHere") {
-			// device.DeviceIP = strings.Split(strings.Split(line, "//")[1], ":")[0]
-			device.WdaReadyChan <- true
-		}
-	}
-
-	if err := cmd.Wait(); err != nil {
-		wdaLogger.LogError("webdriveragent", fmt.Sprintf("Error waiting for WebDriverAgent go-ios command to finish, it errored out or device `%v` was disconnected - %v", device.Device.UDID, err))
-		device.resetLocalDevice()
-	}
-}
-
-// Mount the developer disk images downloading them automatically in /opt/DeveloperDiskImages
-func (device *LocalDevice) mountDeveloperImageIOS() error {
-	basedir := "./devimages"
-
-	var err error
-	path, err := imagemounter.DownloadImageFor(device.GoIOSDeviceEntry, basedir)
-	if err != nil {
-		return err
-	}
-
-	err = imagemounter.MountImage(device.GoIOSDeviceEntry, path)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (device *LocalDevice) pairIOS() error {
-	log.WithFields(log.Fields{
-		"event": "ios_device_setup",
-	}).Debug("Pairing iOS device - " + device.Device.UDID)
-
-	p12, err := os.ReadFile("./config/supervision.p12")
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "ios_device_setup",
-		}).Warn(fmt.Sprintf("Could not read /opt/supervision.p12 file when pairing device with UDID: %s, falling back to unsupervised pairing, err:%s", device.Device.UDID, err))
-		err = ios.Pair(device.GoIOSDeviceEntry)
-		if err != nil {
-			return errors.New("Could not pair successfully, err:" + err.Error())
-		}
-		return nil
-	}
-
-	err = ios.PairSupervised(device.GoIOSDeviceEntry, p12, util.Config.EnvConfig.SupervisionPassword)
-	if err != nil {
-		return errors.New("Could not pair successfully, err:" + err.Error())
-	}
-
-	return nil
 }
